+++ binless/src/fast_binless.cpp
@@ -7,6 +7,8 @@ using namespace Rcpp;
 #include <Eigen/Sparse>
 
 #include "fast_binless.hpp"
+#include "fast_decay.hpp"
+#include "fast_dataframe.hpp"
 #include "GFLLibrary.hpp"
 #include "FusedLassoGaussianEstimator.hpp"
 #include "spline.hpp"
@@ -16,11 +18,11 @@ namespace binless {
 namespace fast {
 
 //here, exposures are log ( sum_i observed / sum_i expected ) with i summed over datasets
-std::vector<double> compute_poisson_lsq_exposures(const FastSignalData& data) {
+std::vector<double> compute_poisson_lsq_exposures(const FastSignalData& data, const DecayEstimator& dec, double pseudocount) {
   //get observed and expected
   std::vector<double> sum_obs(data.get_ndatasets(),0);
   std::vector<double> sum_exp(data.get_ndatasets(),0);
-  auto log_expected = data.get_log_expected();
+  auto log_expected = get_log_expected(data, dec);
   auto observed = data.get_observed();
   //sum them for each dataset
   auto names = data.get_name();
@@ -33,15 +35,15 @@ std::vector<double> compute_poisson_lsq_exposures(const FastSignalData& data) {
   std::vector<double> exposures;
   exposures.reserve(data.get_ndatasets());
   for (unsigned i=0; i<data.get_ndatasets(); ++i) {
-    exposures.push_back(std::log(sum_obs[i]/sum_exp[i]));
+    exposures.push_back(std::log((sum_obs[i]+pseudocount)/sum_exp[i]));
   }
   return exposures;
 }
 
 //one IRLS iteration for exposures, with a poisson model
-std::vector<double> step_exposures(const FastSignalData& data) {
+std::vector<double> step_exposures(const FastSignalData& data, const DecayEstimator& dec) {
   //get residuals
-  ResidualsPair z = get_poisson_residuals(data);
+  ResidualsPair z = get_poisson_residuals(data, dec);
   //average by name
   std::vector<double> exposures(data.get_ndatasets(),0);
   std::vector<double> weightsums(data.get_ndatasets(),0);
@@ -60,11 +62,11 @@ std::vector<double> step_exposures(const FastSignalData& data) {
 }
 
 //here, biases are log ( sum_i observed / sum_i expected ) with i summed over rows/columns
-std::vector<double> compute_poisson_lsq_log_biases(const FastSignalData& data) {
+std::vector<double> compute_poisson_lsq_log_biases(const FastSignalData& data, const DecayEstimator& dec, double pseudocount) {
   //get observed and expected data
   std::vector<double> sum_obs(data.get_nbins(),0);
   std::vector<double> sum_exp(data.get_nbins(),0);
-  auto log_expected = data.get_log_expected();
+  auto log_expected = get_log_expected(data, dec);
   auto observed = data.get_observed();
   //sum them along the rows/columns
   auto dbin1 = data.get_bin1();
@@ -84,7 +86,7 @@ std::vector<double> compute_poisson_lsq_log_biases(const FastSignalData& data) {
   std::vector<double> log_bias;
   log_bias.reserve(data.get_nbins());
   for (unsigned i=0; i<data.get_nbins(); ++i) {
-    log_bias.push_back(std::log(sum_obs[i]/sum_exp[i]));
+    log_bias.push_back(std::log((sum_obs[i]+pseudocount)/sum_exp[i]));
   }
   //center log_bias
   double avg = std::accumulate(log_bias.begin(), log_bias.end(), 0.)/log_bias.size();
@@ -95,9 +97,9 @@ std::vector<double> compute_poisson_lsq_log_biases(const FastSignalData& data) {
 }
 
 //one IRLS iteration for log biases, with a poisson model
-std::vector<double> step_log_biases(const FastSignalData& data) {
+std::vector<double> step_log_biases(const FastSignalData& data, const DecayEstimator& dec) {
   //get residuals
-  ResidualsPair z = get_poisson_residuals(data);
+  ResidualsPair z = get_poisson_residuals(data, dec);
   //sum them along the rows/columns
   auto dbin1 = data.get_bin1();
   auto dbin2 = data.get_bin2();
@@ -130,132 +132,14 @@ std::vector<double> step_log_biases(const FastSignalData& data) {
   for (unsigned i=0; i<data.get_nbins(); ++i) {
     log_biases[i] -= avg;
   }
+  //cap estimates at 3SD from the mean
+  double sq_sum = std::inner_product(log_biases.begin(),log_biases.end(),log_biases.begin(),0.0);
+  double stdev = std::sqrt(sq_sum/log_biases.size());
+  for (unsigned i=0; i<data.get_nbins(); ++i) {
+    if (log_biases[i] > 3*stdev) log_biases[i] = 3*stdev;
+    if (log_biases[i] < -3*stdev) log_biases[i] = -3*stdev;
+  }
   return log_biases;
-}
-
-
-//here, log decay is log ( sum_i observed / sum_i expected ) with i summed over counter diagonals
-std::vector<double> compute_poisson_lsq_log_decay(const FastSignalData& data) {
-  //get observed and expected data
-  std::vector<double> sum_obs(data.get_nbins(),0);
-  std::vector<double> sum_exp(data.get_nbins(),0);
-  auto log_expected = data.get_log_expected();
-  auto observed = data.get_observed();
-  //sum them along the counter diagonals
-  auto dbin1 = data.get_bin1();
-  auto dbin2 = data.get_bin2();
-  for (unsigned i=0; i<data.get_N(); ++i) {
-    unsigned dist = dbin2[i]-dbin1[i];
-    double expected = std::exp(log_expected[i]);
-    sum_obs[dist] += observed[i];
-    sum_exp[dist] += expected;
-  }
-  //compute log_decay
-  std::vector<double> log_decay;
-  log_decay.reserve(data.get_nbins());
-  for (unsigned i=0; i<data.get_nbins(); ++i) {
-    if (sum_obs[i]==0) {
-      Rcpp::Rcout << "counter diag " << i << " is zero!\n";
-      Rcpp::stop(" Aborting...");
-    }
-    log_decay.push_back(std::log(sum_obs[i]/sum_exp[i]));
-  }
-  //center log_decay
-  double avg = std::accumulate(log_decay.begin(), log_decay.end(), 0.)/log_decay.size();
-  for (unsigned i=0; i<data.get_nbins(); ++i) {
-    log_decay[i] -= avg;
-  }
-  return log_decay;
-}
-
-DecaySummary get_decay_summary(const FastSignalData& data) {
-  //get residuals
-  ResidualsPair z = get_poisson_residuals(data);
-  //sum them along the diagonals
-  auto dbin1 = data.get_bin1();
-  auto dbin2 = data.get_bin2();
-  std::vector<double> kappahat(data.get_nbins(),0);
-  std::vector<double> weight(data.get_nbins(),0);
-  for (unsigned i=0; i<data.get_N(); ++i) {
-    unsigned bin1 = dbin1[i]-1; //offset by 1 for vector indexing
-    unsigned bin2 = dbin2[i]-1;
-    kappahat[bin2-bin1] += z.residuals[i]*z.weights[i];
-    weight[bin2-bin1] += z.weights[i];
-  }
-  //add current bias and normalize
-  auto log_decay = data.get_log_decay();
-  std::vector<double> distance;
-  for (unsigned i=0; i<data.get_nbins(); ++i) {
-    kappahat[i] = (weight[i]>0) ? kappahat[i]/weight[i] : 0;
-    kappahat[i] += log_decay[i];
-    distance.push_back(i+1);
-  }
-  return DecaySummary{distance,kappahat,weight};
-}
-
-DecayFit pointwise_log_decay_fit(const DecaySummary& dec) {
-  //forbid increase past first diagonal
-  std::vector<double> log_decay = dec.kappahat;
-  unsigned N = log_decay.size();
-  for (unsigned i=0; i<N; ++i) {
-    if (i>=2 && log_decay[i] > log_decay[i-1]) log_decay[i] = log_decay[i-1];
-  }
-  //compute weighted mean
-  double avg=0;
-  double wsum=0;
-  for (unsigned i=0; i<N; ++i) {
-    avg += log_decay[i]*dec.weight[i];
-    wsum += dec.weight[i];
-  }
-  avg = avg / wsum;
-  //no smoothing, subtract average and return
-  for (unsigned i=0; i<N; ++i) {
-    log_decay[i] -= avg;
-  }
-  return DecayFit{log_decay,dec,-1};
-}
-
-DecayFit spline_log_decay_fit(const DecaySummary& dec, double tol_val, unsigned Kdiag, unsigned max_iter, double sigma) {
-  //extract data
-  const Eigen::Map<const Eigen::VectorXd> y(dec.kappahat.data(),dec.kappahat.size());
-  const Eigen::Map<const Eigen::VectorXd> w(dec.weight.data(),dec.weight.size());
-  const Eigen::VectorXd S = w.array().inverse().sqrt().matrix();
-  //X: build spline base on log distance
-  const Eigen::Map<const Eigen::ArrayXd> distance(dec.distance.data(), dec.distance.size());
-  const Eigen::VectorXd log_distance = distance.log().matrix();
-  const Eigen::SparseMatrix<double> X = generate_spline_base(log_distance, Kdiag);
-  //D: build difference matrix
-  const Eigen::SparseMatrix<double> D = second_order_difference_matrix(Kdiag);
-  //C: build constraint matrix to forbid increase
-  const Eigen::SparseMatrix<double> Cin = - first_order_difference_matrix(Kdiag);
-  
-  //iteratively fit GAM on decay and estimate lambda
-  GeneralizedAdditiveModel gam(y,S,X,D,sigma);
-  gam.set_inequality_constraints(Cin);
-  gam.optimize(max_iter,tol_val);
-  //Rcpp::Rcout << "gam converged: " << gam.has_converged() << "\n";
-  Eigen::VectorXd log_decay = gam.get_mean();
-  const double lambda = gam.get_lambda();
-  
-  //compute weighted mean
-  double avg = w.dot(log_decay)/w.sum();
-  //subtract average and return
-  log_decay.array() -= avg;
-  return DecayFit{std::vector<double>(log_decay.data(),log_decay.data()+log_decay.rows()),dec,lambda};
-}
-
-//one IRLS iteration for log decay, with a poisson model
-std::vector<double> step_log_decay(const FastSignalData& data, double tol_val) {
-  //compute summary statistics for decay
-  DecaySummary dec = get_decay_summary(data);
-  //infer new decay from summaries
-  //DecayFit fit = pointwise_log_decay_fit(dec);
-  DecayFit fit = spline_log_decay_fit(dec, tol_val);
-  /*Rcpp::Rcout << "distance log_decay kappahat weight\n";
-  for (unsigned i=0; i<fit.log_decay.size(); ++i)
-    Rcpp::Rcout << fit.dec.distance[i] << " " << fit.log_decay[i] << " "
-                << fit.dec.kappahat[i] << " " << fit.dec.weight[i] << "\n";*/
-  return fit.log_decay;
 }
 
 PrecisionPair get_precision(const std::vector<double>& weights, const std::vector<double>& weights_old) {
@@ -325,9 +209,15 @@ List binless(const DataFrame obs, unsigned nbins, double lam2, unsigned ngibbs, 
   //initialize return values, exposures and fused lasso optimizer
   Rcpp::Rcout << "init\n";
   FastSignalData out(obs, nbins);
-  out.set_exposures(compute_poisson_lsq_exposures(out));
-  out.set_log_decay(compute_poisson_lsq_log_decay(out));
-  out.set_log_biases(compute_poisson_lsq_log_biases(out));
+  DecayConfig conf(tol_val);
+  DecayEstimator dec(out, conf);
+  //
+  out.set_exposures(compute_poisson_lsq_exposures(out, dec));
+  //
+  dec.set_poisson_lsq_summary(out);
+  dec.update_params();
+  //
+  out.set_log_biases(compute_poisson_lsq_log_biases(out, dec));
   double current_tol_val = 1.;
   std::vector<FusedLassoGaussianEstimator<GFLLibrary> > flos(out.get_ndatasets(),
                                                              FusedLassoGaussianEstimator<GFLLibrary>(nbins, current_tol_val/20.));
@@ -337,21 +227,20 @@ List binless(const DataFrame obs, unsigned nbins, double lam2, unsigned ngibbs, 
     Rcpp::Rcout << "step " << step;
     std::vector<double> expected,old_expected;
     //compute biases
-    auto biases = step_log_biases(out);
+    auto biases = step_log_biases(out, dec);
     out.set_log_biases(biases);
     //compute decay
-    auto decay = step_log_decay(out, tol_val);
-    if (step <= bg_steps) old_expected = out.get_log_expected();
-    out.set_log_decay(decay);
-    if (step <= bg_steps) expected = out.get_log_expected();
+    if (step <= bg_steps) old_expected = get_log_expected(out, dec);
+    dec.step_irls(out);
+    if (step <= bg_steps) expected = get_log_expected(out, dec);
     //compute signal
     if (step > bg_steps) {
-      old_expected = out.get_log_expected();
-      auto signal = step_signal(out, flos, lam2);
+      old_expected = get_log_expected(out, dec);
+      auto signal = step_signal(out, dec, flos, lam2);
       out.set_log_signal(signal.beta);
       out.set_signal_phihat(signal.phihat);
       out.set_signal_weights(signal.weights);
-      expected = out.get_log_expected();
+      expected = get_log_expected(out, dec);
     }
     //compute precision and convergence
     auto precision = get_precision(expected,old_expected);
@@ -373,9 +262,9 @@ List binless(const DataFrame obs, unsigned nbins, double lam2, unsigned ngibbs, 
         out.set_log_signal(adjust);
       }
       //compute exposures
-      auto exposures = step_exposures(out);
+      auto exposures = step_exposures(out, dec);
       out.set_exposures(exposures);
-      //diagnostics.push_back(out.get_as_dataframe());
+      //diagnostics.push_back(get_as_dataframe(out,dec));
       if (converged) {
         Rcpp::Rcout << "converged\n";
         break;
@@ -383,8 +272,8 @@ List binless(const DataFrame obs, unsigned nbins, double lam2, unsigned ngibbs, 
   }
   Rcpp::Rcout << "done\n";
   //finalize and return
-  return Rcpp::List::create(_["mat"]=out.get_as_dataframe(), _["log_biases"]=out.get_log_biases(),
-                            _["log_decay"]=out.get_log_decay(), _["exposures"]=out.get_exposures(),
+  return Rcpp::List::create(_["mat"]=get_as_dataframe(out,dec), _["log_biases"]=out.get_log_biases(),
+                            _["beta_diag"]=dec.get_beta(), _["exposures"]=out.get_exposures(),
                             //_["diagnostics"]=diagnostics,
                             _["nbins"]=nbins);
 }
@@ -398,8 +287,12 @@ Rcpp::List binless_eval_cv(const List obs, const NumericVector lam2, unsigned gr
   if (out.get_ndatasets() != 1) Rcpp::stop("Provide only 1 dataset!");
   auto signal_ori = Rcpp::as<std::vector<double> >(mat["signal"]);
   out.set_log_signal(signal_ori); //fills-in phi_ref and delta
-  auto log_decay = Rcpp::as<std::vector<double> >(obs["log_decay"]);
-  out.set_log_decay(log_decay);
+  //
+  DecayConfig conf(tol_val);
+  DecayEstimator dec(out, conf);
+  auto beta_diag = Rcpp::as<Eigen::VectorXd >(obs["beta_diag"]);
+  dec.set_beta(beta_diag);
+  //
   auto log_biases = Rcpp::as<std::vector<double> >(obs["log_biases"]);
   out.set_log_biases(log_biases);
   auto exposures = Rcpp::as<std::vector<double> >(obs["exposures"]);
@@ -413,11 +306,11 @@ Rcpp::List binless_eval_cv(const List obs, const NumericVector lam2, unsigned gr
     Rcpp::checkUserInterrupt();
     Rcpp::Rcout << "lambda2= " << lam2[i] << "\n";
     out.set_log_signal(signal_ori); //fills-in phi_ref and delta
-    auto signal = step_signal(out, flos, lam2[i], group);
+    auto signal = step_signal(out, dec, flos, lam2[i], group);
     out.set_log_signal(signal.beta);
     out.set_signal_phihat(signal.phihat);
     out.set_signal_weights(signal.weights);
-    diagnostics.push_back(out.get_as_dataframe());
+    diagnostics.push_back(get_as_dataframe(out,dec));
   }
   //finalize and return
   Rcpp::Rcout << "done\n";
@@ -432,8 +325,12 @@ Rcpp::DataFrame binless_difference(const List obs, double lam2, unsigned ref, do
   FastDifferenceData out(mat, nbins, ref);
   auto signal = Rcpp::as<std::vector<double> >(mat["log_signal"]);
   out.set_log_signal(signal); //fills-in phi_ref and delta
-  auto log_decay = Rcpp::as<std::vector<double> >(obs["log_decay"]);
-  out.set_log_decay(log_decay);
+  //
+  DecayConfig conf(tol_val);
+  DecayEstimator dec(out, conf);
+  auto beta_diag = Rcpp::as<Eigen::VectorXd >(obs["beta_diag"]);
+  dec.set_beta(beta_diag);
+  //
   auto log_biases = Rcpp::as<std::vector<double> >(obs["log_biases"]);
   out.set_log_biases(log_biases);
   auto exposures = Rcpp::as<std::vector<double> >(obs["exposures"]);
@@ -443,14 +340,14 @@ Rcpp::DataFrame binless_difference(const List obs, double lam2, unsigned ref, do
                                                              FusedLassoGaussianEstimator<GFLLibrary>(nbins, converge));
   //compute differences
   Rcpp::Rcout << "compute\n";
-  auto diff = step_difference(out, flos, lam2, ref);
+  auto diff = step_difference(out, dec, flos, lam2, ref);
   out.set_log_difference(diff.delta);
   out.set_deltahat(diff.deltahat);
   out.set_difference_weights(diff.weights);
   out.set_phi_ref(diff.phi_ref);
   //finalize and return
   Rcpp::Rcout << "done\n";
-  return out.get_as_dataframe();
+  return get_as_dataframe(out);
 }
 
 }