+++ binless/src/util.cpp
@@ -53,39 +53,50 @@ double get_maximum_admissible_weight(const Rcpp::NumericVector& w, double percen
     return sw[i];
 }
 
-std::map<double,unsigned> make_even_bins(double lower, double upper, unsigned Nbins) {
+BinDesign make_even_bins(double lower, double upper, unsigned Nbins) {
   double size = upper - lower;
   std::map<double,unsigned> ret;
+  std::vector<double> begins, ends;
   for (unsigned i=0; i<=Nbins; ++i) { //size of map is Nbins+1
     double begin = lower + i*size/double(Nbins);
     ret[begin]=i;
+    if (i<Nbins) begins.push_back(begin);
+    if (i>0) ends.push_back(begin);
   }
-  return ret;
+  return BinDesign{Nbins,ret,begins,ends};
 }
 
-Eigen::SparseMatrix<unsigned> bin_data(const Eigen::VectorXd& data, const std::map<double,unsigned>& begin2bin) {
-  unsigned Nbins = begin2bin.size()-1;
+Eigen::SparseMatrix<double, Eigen::RowMajor> bin_data(const Eigen::VectorXd& data, const BinDesign& design, bool drop) {
   unsigned Ndata = data.rows();
-  std::vector<Eigen::Triplet<unsigned> > tripletList;
+  std::vector<Eigen::Triplet<double> > tripletList;
   tripletList.reserve(Ndata);
   for (unsigned i=0; i<Ndata; ++i) {
-    auto it = begin2bin.upper_bound(data(i));
-    if (it == begin2bin.end()) it = begin2bin.lower_bound(data(i));
+    auto it = design.map.upper_bound(data(i));
+    if (it == design.map.end()) it = design.map.lower_bound(data(i));
     unsigned j = it->second - 1; //upper_bound returns pointer to next bin
-    tripletList.push_back(Eigen::Triplet<unsigned>(j,i,1));
+    tripletList.push_back(Eigen::Triplet<double>(j,i,1.));
   }
-  Eigen::SparseMatrix<unsigned> mat(Nbins,Ndata);
+  Eigen::SparseMatrix<double, Eigen::RowMajor> mat(design.Nbins,Ndata);
   mat.setFromTriplets(tripletList.begin(), tripletList.end());
-  mat.makeCompressed();
-  return mat;
+  if (drop) {
+    Eigen::SparseMatrix<double, Eigen::RowMajor> ret(design.Nbins,Ndata);
+    unsigned Nrow=0;
+    for (unsigned i=0; i<design.Nbins; ++i) {
+      if (mat.row(i).sum()>0) ret.row(Nrow++) = mat.row(i);
+    }
+    ret.conservativeResize(Nrow,Ndata);
+    ret.makeCompressed();
+    return ret;
+  } else {
+    mat.makeCompressed();
+    return mat;
+  }
 }
 
-Eigen::SparseMatrix<unsigned> bin_data_evenly(const Eigen::VectorXd& data, unsigned Nbins) {
-  auto begin2bin = make_even_bins(data.minCoeff(),data.maxCoeff(),Nbins);
-  return bin_data(data,begin2bin);
+Eigen::SparseMatrix<double, Eigen::RowMajor> bin_data_evenly(const Eigen::VectorXd& data, unsigned Nbins, bool drop) {
+  auto design = make_even_bins(data.minCoeff(),data.maxCoeff(),Nbins);
+  return bin_data(data,design,drop);
 }
-
-