+++ fast_binless.R
@@ -17,16 +17,6 @@ mat=binless:::bin_data(cs,resolution=5000)
 #write.table(mat,file = "example/rao_HiCall_FOXP1ext_2.3M_mat_5kb.dat", quote = T, row.names = F)
 
 
-#In fast binless, you must ensure no counter diagonal nor row/column is completely zero
-#if there are not too many, you can add 1 to the observed counts
-#also, because of the simplified decay, the very last bin cannot be zero in all datasets
-#empty rows:
-rbind(mat[,.(bin=bin1,observed)],mat[,.(bin=bin2,observed)])[,.(sum(observed)),by=bin][V1==0]
-#empty counter diagonals:
-mat[,.(d=unclass(bin2)-unclass(bin1),observed)][,sum(observed),by=d][V1==0]
-#last bin:
-mat[unclass(bin1)==1&unclass(bin2)==max(unclass(bin2))]
-
 #The fast binless algorithm computes a binless normalization without estimating
 #the fusion penalty and the significance threshold.
 #We do a maximum of nouter steps (or less if the relative precision is lower than tol_val)
@@ -54,7 +44,7 @@ ggplot(data.table(bin=1:nlevels(mat[,bin1]),log_biases=out$log_biases))+geom_poi
 #biases matrix
 plot_binless_matrix(a, upper="log_biases", lower="log_biases", trans="identity")
 #fitted decay
-ggplot(unique(mat[,.(distance,log_decay=out$log_decay)]))+geom_line(aes(distance,log_decay))#+scale_x_log10()
+ggplot(unique(a[,.(distance,log_decay)]))+geom_line(aes(distance,log_decay))+scale_x_log10()
 #decay matrix
 plot_binless_matrix(a, upper="log_decay", lower="log_decay", trans="identity")
 #signal matrix ( = what is different from the background)